feature: refactor for optimization
